New book, audible banner update, nav and footer update.
Added information for new book such as a related section in the index, links in the footers and navs, and finally a page of its own (templated off of the book 3 page along with content from the bio page).

changed Notice on the Warriors Stone page to reflect the release on audible.
Added link to audible image
added audible banner to book 4 page,